feat(pace-117): fixes feat/pace-119
