+++ app/src/main/kotlin/com/getstrm/pace/api/ProcessingPlatformsApi.kt
@@ -3,14 +3,14 @@ package com.getstrm.pace.api
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.ProcessingPlatform
 import build.buf.gen.getstrm.pace.api.processing_platforms.v1alpha.*
 import com.getstrm.pace.processing_platforms.ProcessingPlatformClient
-import com.getstrm.pace.service.DataPolicyService
+import com.getstrm.pace.service.LineageService
 import com.getstrm.pace.service.ProcessingPlatformsService
 import net.devh.boot.grpc.server.service.GrpcService
 
 @GrpcService
 class ProcessingPlatformsApi(
     private val processingPlatformsService: ProcessingPlatformsService,
-    private val dataPolicyService: DataPolicyService,
+    private val lineageService: LineageService,
 ) : ProcessingPlatformsServiceGrpcKt.ProcessingPlatformsServiceCoroutineImplBase() {
 
     override suspend fun listProcessingPlatforms(
@@ -57,14 +57,6 @@ class ProcessingPlatformsApi(
         processingPlatformsService.getBlueprintPolicy(request)
 
     override suspend fun getLineage(request: GetLineageRequest): GetLineageResponse {
-        val lineage = processingPlatformsService.getLineage(request)
-        val withDataPolicyInfo =
-            dataPolicyService.determineLineageDataPolicies(lineage.lineageSummary)
-        val enriched =
-            with(lineage.toBuilder()) {
-                lineageSummary = withDataPolicyInfo
-                build()
-            }
-        return enriched
+        return lineageService.getLineage(request)
     }
 }
